// src/features/auth/presentation/types/AuthFormTypes.ts
export interface LoginFormValues {
  email: string;
  password: string;
}

export interface SignupStep1Values {
  email: string;
  password: string;
  passwordConfirm: string;
  nickname: string;
}

export interface SignupStep2Values {
  name: string;
  birthday: string;
  age: string;
  job: string;
}